refactor: DTO 및 type 적용
