+++ backend/src/auth/auth.service.spec.ts
@@ -10,12 +10,13 @@ import { mockUser } from '../fixtures/users.fixture';
 import { Users } from '../users/users.entity';
 import { UsersRepository } from '../users/users.repository';
 import { UsersService } from '../users/users.service';
-import { OauthBody } from './auth.controller';
-import { AuthService, OauthData } from './auth.service';
+import { AuthService } from './auth.service';
+import { OAuthAuthorizeDTO } from './dtos/oauth.dto';
 import { GoogleService } from './oauth/google.service';
 import { KakaoService } from './oauth/kakao.service';
 import { NaverService } from './oauth/naver.service';
 import { TokenService } from './token/token.service';
+import { OAUTH_PROVIDERS, OauthData } from './types/auth.type';
 
 const context = describe;
 
@@ -96,16 +97,15 @@ describe('AuthService', () => {
     });
 
     const authorizeCode = 'authorizeCode';
-    const providerList = ['google', 'kakao', 'naver'];
 
     describe('login', () => {
         context('사용자가 존재하면', () => {
             for (let i = 0; i < 3; i++) {
-                const provider = providerList[i];
+                const provider = OAUTH_PROVIDERS[i];
                 it(`${provider} 로그인 후 access token과 refresh token을 반환해야 한다.`, async () => {
                     jest.spyOn(usersService, 'updateAndFindOne').mockResolvedValue({ id: 1 } as Users);
 
-                    const result = await service.login({ authorizeCode, provider } as OauthBody);
+                    const result = await service.login({ authorizeCode, provider } as OAuthAuthorizeDTO);
 
                     expect(result).toEqual({
                         accessToken: mockUser.refreshToken,
@@ -117,11 +117,11 @@ describe('AuthService', () => {
 
         context('사용자가 존재하지 않으면', () => {
             for (let i = 0; i < 3; i++) {
-                const provider = providerList[i];
+                const provider = OAUTH_PROVIDERS[i];
                 it(`${provider} 로그인 후 oauth data를 반환해야 한다.`, async () => {
                     jest.spyOn(usersService, 'updateAndFindOne').mockRejectedValue(new NotFoundException());
 
-                    const result = await service.login({ authorizeCode, provider } as OauthBody);
+                    const result = await service.login({ authorizeCode, provider } as OAuthAuthorizeDTO);
 
                     expect(result).toEqual({
                         oauthAccessToken: mockUser.oauthAccessToken,
@@ -136,7 +136,7 @@ describe('AuthService', () => {
     describe('signup', () => {
         context('사용자가 존재하지 않으면', () => {
             for (let i = 0; i < 3; i++) {
-                const provider = providerList[i];
+                const provider = OAUTH_PROVIDERS[i];
                 it(`${provider} 회원가입 후 access token과 refresh token을 반환해야 한다.`, async () => {
                     jest.spyOn(usersService, 'createIfNotExists').mockResolvedValue({ id: 1 } as Users);
